update/fix: Dynamic routing for existing incidents route
